[#349] Clean up React Files WIP
